fix(instance-audit): unblock type and complexity gates
