+++ apps/sva-studio-react/src/routes/admin/instances/-instances-shared.test.tsx
@@ -328,6 +328,7 @@ describe('instances shared helpers', () => {
         clientExists: true,
         tenantAdminExists: true,
         tenantAdminHasSystemAdmin: true,
+        systemAdminRoleExists: true,
         redirectUrisMatch: true,
         logoutUrisMatch: true,
         webOriginsMatch: true,
@@ -379,6 +380,7 @@ describe('instances shared helpers', () => {
           tenantAdminClientExists: false,
           tenantAdminExists: false,
           tenantAdminHasSystemAdmin: false,
+          systemAdminRoleExists: false,
           redirectUrisMatch: true,
           logoutUrisMatch: true,
           webOriginsMatch: true,
@@ -434,6 +436,7 @@ describe('instances shared helpers', () => {
           tenantAdminClientExists: false,
           tenantAdminExists: true,
           tenantAdminHasSystemAdmin: true,
+          systemAdminRoleExists: true,
           redirectUrisMatch: true,
           logoutUrisMatch: true,
           webOriginsMatch: true,
@@ -644,6 +647,7 @@ describe('instances shared helpers', () => {
           tenantAdminClientExists: true,
           tenantAdminExists: true,
           tenantAdminHasSystemAdmin: true,
+          systemAdminRoleExists: true,
           redirectUrisMatch: true,
           logoutUrisMatch: true,
           webOriginsMatch: true,
@@ -749,6 +753,7 @@ describe('instances shared helpers', () => {
           tenantAdminClientExists: true,
           tenantAdminExists: true,
           tenantAdminHasSystemAdmin: true,
+          systemAdminRoleExists: true,
           redirectUrisMatch: true,
           logoutUrisMatch: true,
           webOriginsMatch: true,
@@ -783,6 +788,7 @@ describe('instances shared helpers', () => {
         tenantAdminClientExists: true,
         tenantAdminExists: true,
         tenantAdminHasSystemAdmin: true,
+        systemAdminRoleExists: true,
         redirectUrisMatch: true,
         logoutUrisMatch: true,
         webOriginsMatch: true,
@@ -1063,6 +1069,7 @@ describe('instances shared helpers', () => {
           tenantAdminClientExists: false,
           tenantAdminExists: false,
           tenantAdminHasSystemAdmin: false,
+          systemAdminRoleExists: false,
           redirectUrisMatch: false,
           logoutUrisMatch: false,
           webOriginsMatch: false,
@@ -1133,6 +1140,7 @@ describe('instances shared helpers', () => {
           tenantAdminClientExists: true,
           tenantAdminExists: true,
           tenantAdminHasSystemAdmin: true,
+          systemAdminRoleExists: true,
           redirectUrisMatch: true,
           logoutUrisMatch: true,
           webOriginsMatch: true,
@@ -1207,6 +1215,7 @@ describe('instances shared helpers', () => {
           tenantAdminClientExists: true,
           tenantAdminExists: true,
           tenantAdminHasSystemAdmin: true,
+          systemAdminRoleExists: true,
           redirectUrisMatch: true,
           logoutUrisMatch: false,
           webOriginsMatch: true,
@@ -1258,6 +1267,7 @@ describe('instances shared helpers', () => {
           tenantAdminClientExists: true,
           tenantAdminExists: true,
           tenantAdminHasSystemAdmin: true,
+          systemAdminRoleExists: true,
           redirectUrisMatch: true,
           logoutUrisMatch: true,
           webOriginsMatch: true,
